+++ api/blogQueries.go
@@ -196,14 +196,14 @@ var blogQueryFields = graphql.Fields{
 					} else {
 						blogData["created"] = objectidTimestamp(id).Unix()
 					}
-					updatedInt, ok := blogData["updated"].(int32)
+					updatedInt, ok := blogData["updated"].(int64)
 					if !ok {
 						return nil, errors.New("cannot cast updated time to int")
 					}
 					if formatDate {
-						blogData["updated"] = intTimestamp(int64(updatedInt)).Format(dateFormat)
+						blogData["updated"] = intTimestamp(updatedInt).Format(dateFormat)
 					} else {
-						blogData["updated"] = intTimestamp(int64(updatedInt)).Unix()
+						blogData["updated"] = intTimestamp(updatedInt).Unix()
 					}
 					blogData["id"] = id.Hex()
 					delete(blogData, "_id")
@@ -331,14 +331,14 @@ var blogQueryFields = graphql.Fields{
 				} else {
 					blogData["created"] = objectidTimestamp(id).Unix()
 				}
-				updatedInt, ok := blogData["updated"].(int32)
+				updatedInt, ok := blogData["updated"].(int64)
 				if !ok {
 					return nil, errors.New("cannot cast updated time to int")
 				}
 				if formatDate {
-					blogData["updated"] = intTimestamp(int64(updatedInt)).Format(dateFormat)
+					blogData["updated"] = intTimestamp(updatedInt).Format(dateFormat)
 				} else {
-					blogData["updated"] = intTimestamp(int64(updatedInt)).Unix()
+					blogData["updated"] = intTimestamp(updatedInt).Unix()
 				}
 				blogData["id"] = idstring
 				if blogData["tileimage"] != nil {
